feat: Added custom scopes to OAuthToken implementation.
