+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>io.mymetaverse.sdk</groupId>
     <artifactId>java-sdk</artifactId>
-    <version>2.3.1</version>
+    <version>2.5.0</version>
 
     <properties>
         <maven.compiler.source>8</maven.compiler.source>
